daily_data_kap_2/requirements.txt
google-genai
python-dotenv
requests
beautifulsoup4
pandas
toml
tweepy
Pillow
chromadb
sentence-transformers
